Mostly working state

+++ config/apps/iterm/iterm.sh
@@ -1,10 +1,14 @@
 #!/usr/bin/env bash
 
 # iTerm Setup Script
-osascript -e 'tell application "iTerm" to activate'
-sleep 5
 osascript -e 'tell application "iTerm" to quit'
 
 # Symlink Preferences
-mv ~/Library/Preferences/com.googlecode.iterm2.plist ~/Library/Preferences/com.googlecode.iterm2backup.plist
-ln -s ./com.googlecode.iterm2.plist ~/Library/Preferences/com.googlecode.iterm2.plist
+if [ -f ~/Library/Preferences/com.googlecode.iterm2.plist ]; then
+  mv ~/Library/Preferences/com.googlecode.iterm2.plist ~/Library/Preferences/com.googlecode.iterm2backup.plist
+fi
+
+defaults delete com.googlecode.iterm2 2>&1 /dev/null
+
+ln -s ~/dotfiles/config/apps/iterm/com.googlecode.iterm2.plist ~/Library/Preferences/com.googlecode.iterm2.plist
+defaults read com.googlecode.iterm2 2>&1 /dev/null
