+++ config/apps/vscode/vscode.sh
@@ -4,7 +4,7 @@
 
 # Open/Close VSCode to generate preferences for the first time
 osascript -e 'tell application "Visual Studio Code" to activate'
-sleep 5
+sleep 2
 osascript -e 'tell application "Visual Studio Code" to quit'
 
 # Reload ~/.zshrc to ensure we have access to `code` command in path
@@ -29,8 +29,11 @@ osascript -e 'tell application "Visual Studio Code" to quit'
 # Symlink VSCode preferences
 mv ~/Library/Application\ Support/Code/User/settings.json ~/Library/Application\ Support/Code/User/settings_backup.json
 mv ~/Library/Application\ Support/Code/User/keybindings.json ~/Library/Application\ Support/Code/User/keybindings_backup.json
-mv ~/Library/Application\ Support/Code/User/snippets/ ~/Library/Application\ Support/Code/User/snippets_backup/
+
+if [ -f ~/Library/Application\ Support/Code/User/snippets/ ]; then  
+  mv ~/Library/Application\ Support/Code/User/snippets/ ~/Library/Application\ Support/Code/User/snippets_backup/
+fi
 
 ln -s ~/dotfiles/config/apps/vscode/settings.json ~/Library/Application\ Support/Code/User/settings.json
 ln -s ~/dotfiles/config/apps/vscode/keybindings.json ~/Library/Application\ Support/Code/User/keybindings.json
-ln -s ~/dotfiles/config/apps/vscode/snippets/ ~/Library/Application\ Support/Code/User/snippets/
+# ln -s ~/dotfiles/config/apps/vscode/snippets/ ~/Library/Application\ Support/Code/User/snippets/
